Refactor duplicated methods in ScanValidators
