+++ app/javascript/shared/components/tubeScanValidators.spec.js
@@ -1,18 +1,4 @@
-import { checkDuplicates, checkState, aggregate } from 'shared/components/tubeScanValidators'
-
-describe('aggregate', () => {
-  const validFunction = (_) => { return { valid: true, message: 'Good' } }
-  const invalidFunction = (_) => { return { valid: false, message: 'Bad' } }
-
-  it('is valid if all functions are valid', () => {
-    expect(aggregate([validFunction, validFunction], {})).toEqual({ valid: true, message: 'Good' })
-  })
-
-  it('is invalid if any functions are invalid', () => {
-    expect(aggregate([validFunction, invalidFunction], {})).toEqual({ valid: false, message: 'Bad' })
-    expect(aggregate([invalidFunction, validFunction], {})).toEqual({ valid: false, message: 'Bad' })
-  })
-})
+import { checkDuplicates, checkState } from 'shared/components/tubeScanValidators'
 
 describe('checkDuplicates', () => {
   it('passes if it has distinct tubes', () => {
